Update to be compatible with Pylint 2.1.1 and Python 3
When checking calls to the builtin open function without manager we now
have to check that the module of the 'open' function is '_io'.

We remove the 'file' definition and its corresponding 'builtin-name-used' expectation in 'builtin-name-used.py' since 'file'is no longer a builtin in Python 3.

We replace the Python 2 'print' statements by calls to the Python 3
'print' function.

The Tokenize library used by Pylint now adds an 'ENCODING' token to the
start of the token list when tokenizing a file. We have to ignore this
token since it isn't representing any line of code.